project reviewed

+++ database/migrations/2023_05_25_153452_create_trains_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->dateTime('estimated_departure');
             $table->dateTime('estimated_arrival');
             $table->string('train_code', 10);
-            $table->smallInteger('N_coach')->unsigned();
+            $table->tinyInteger('N_coach')->unsigned();
             $table->boolean('on_time')->default(true);
             $table->boolean('cancelled')->default(false);
             $table->timestamps();
